+++ packages/core/src/index.ts
@@ -1,10 +1,10 @@
 import { useVirtualRowContext } from "./table/rows/virtual_row_context";
 import { ReactTanstackTableUi } from "./table/table";
 import { iterateOverColumns } from "./iterate_over_columns";
-import { Skin } from "./skin";
+import type { Skin } from "./skin";
 import { useTableContext } from "./table/table_context";
 import { useColContext } from "./table/cols/col_context";
-import { VirtualHeader } from "./table/cols/virtual_header/types";
+import type { VirtualHeader } from "./table/cols/virtual_header/types";
 import { useTableCssVars } from "./skin";
 import {
   defaultSkin,
@@ -14,6 +14,8 @@ import {
 import { useVirtualHeader } from "./table/cols/virtual_header/context";
 import { useRow } from "./table/rows/row_context";
 import { useMeasureHeader } from "./use_measure_header";
+import { decorateColumnHelper } from "./decorate_column_helper";
+import type { HeaderDecorator, CellDecorator } from "./decorate_column_helper";
 export {
   ReactTanstackTableUi,
   useVirtualRowContext,
@@ -27,5 +29,6 @@ export {
   lightModeVars,
   useVirtualHeader,
   useMeasureHeader,
+  decorateColumnHelper,
 };
-export type { Skin, VirtualHeader };
+export type { Skin, VirtualHeader, HeaderDecorator, CellDecorator };