+++ packages/skin-tailwind/src/CellCode.tsx
@@ -1,11 +1,15 @@
 import React from "react";
 
-
-export const CellCode: React.FC<{ children: React.ReactNode; }> = ({
-  children,
-}) => {
+export const CellCode: React.FC<{
+  children?: React.ReactNode;
+  code?: string;
+}> = ({ children, code }) => {
+  const props = code ? { dangerouslySetInnerHTML: { __html: code } } : {};
   return (
-    <div className="font-mono text-sm text-gray-800 dark:text-gray-400">
+    <div
+      className="font-mono text-sm text-gray-800 dark:text-gray-400"
+      {...props}
+    >
       {children}
     </div>
   );